Remove the need to click the New button
Previously when a new calculation was required you had to click
the 'New' button. Remove that requirement, you can now just type
in a new amount in either the gross or net fields and hit enter.

This not only enhances the user experience but resulted in less
and simpler code. Win, Win.