Remove the Calculate button
Firstly, this was causing a segfault by passing NULL into cb_calculate
via the gpointer data argument, which was tripping up the strcmp checks.

In an attempt to fix this, it became more apparent that is actually
superfluous and not really clear how to handle the calculate button and
the ability to hit enter in the gross and net fields. Hitting enter in
those fields tells cb_calculate which thing to calculate, if you hit the
Calculate button then it's not clear which one to calculate if both
fields contain data.

Seemed more clear just to remove the thing.