Tidy up the help window
Create the help window at a reasonable initial size, but allow it to be
resized.

By setting the GTK_WRAP_WORD property of the text view we can get rid of
all the \n's that were used to break up the lines in the paragraphs.

Now the help text will be flowed properly as the window size is changed.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>